Crate hopr_chain_config

Re-exports

pub use errors::HoprConfigError;

pub use errors::Result;

Modules

errors

Structs

Addresses

ChainNetworkConfig

ChainOptions

Holds all information we need about the blockchain network the client is going to use

Network

Holds all information about the protocol network to be used by the client

ProtocolsConfig

The entire on-chain protocol configuration containing the information about usable networks and chains.

Enums

EnvironmentType

Types of HOPR network environments.